% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HailwoodHorrobinFuns.R
\name{wemcHH}
\alias{wemcHH}
\title{Wood EMC using Hailwood-Horrobin methods}
\usage{
wemcHH(RH, T, Tunit)
}
\arguments{
\item{RH}{The relative humidity of the air in percent.}

\item{T}{The temperature.}

\item{Tunit}{The unit used for temperature, either
degree celsius ("C") or fahrenheit ("F").}
}
\value{
The EMC value of wood for the given temperature
(T) and relative humidity (RH)
}
\description{
Function to calculate the EMC of wood
using Hailwood-Horrobin methods. Standard values
all around with no distinction between softwood
and hardwood species.
}
\examples{
library(wemcR)

# Use only a combination of one temperature and humidity
T_1 <- 20
Tunit_1 <- "C"
RH_1 <- 65

EMC_1 <- wemcHH(RH_1, T_1, Tunit_1)
}
